Changed Version in BulkResponseItem to int64

+++ bulk.go
@@ -304,7 +304,7 @@ type BulkResponseItem struct {
 	Index   string        `json:"_index,omitempty"`
 	Type    string        `json:"_type,omitempty"`
 	Id      string        `json:"_id,omitempty"`
-	Version int           `json:"_version,omitempty"`
+	Version int64         `json:"_version,omitempty"`
 	Status  int           `json:"status,omitempty"`
 	Found   bool          `json:"found,omitempty"`
 	Error   *ErrorDetails `json:"error,omitempty"`
